LPS-109249 Extract LanguageListItem to a new module
